Fix most warnings

+++ src/caniot.c
@@ -247,7 +247,7 @@ int caniot_explain_frame_str(const struct caniot_frame *frame, char *buf, size_t
 	size_t total = 0U;
 
 	ret = caniot_explain_id_str(frame->id, buf, len);
-	if (ret > len || ret < 0) {
+	if (ret > (int) len || ret < 0) {
 		return ret;
 	}
 
@@ -264,7 +264,7 @@ int caniot_explain_frame_str(const struct caniot_frame *frame, char *buf, size_t
 		   (frame->id.type == CANIOT_FRAME_TYPE_COMMAND)) {
 		ret = snprintf(buf, len, "ep : %s", 
 			       get_endpoint_str(frame->id.endpoint));
-		if (ret > len || ret < 0) {
+		if (ret > (int) len || ret < 0) {
 			return ret;
 		}
 		total += ret;
@@ -273,7 +273,7 @@ int caniot_explain_frame_str(const struct caniot_frame *frame, char *buf, size_t
 
 		for (int i = 0; i < frame->len; i++) {
 			ret = snprintf(buf, len, " %02hhx", (uint8_t)frame->buf[i]);
-			if (ret > len || ret < 0) {
+			if (ret > (int) len || ret < 0) {
 				return ret;
 			}
 			total += ret;
@@ -283,7 +283,7 @@ int caniot_explain_frame_str(const struct caniot_frame *frame, char *buf, size_t
 	} else {
 		ret = snprintf(buf, len, "LEN = %d, key = %02x val = %04x",
 			       frame->len, frame->attr.key, frame->attr.val);
-		if (ret > len || ret < 0) {
+		if (ret > (int) len || ret < 0) {
 			return ret;
 		}
 		total += ret;
